feat: 여행 초기 계획 데이터 포맷팅 함수 생성
