+++ src/apis/useArticle/fetch.ts
@@ -1,10 +1,10 @@
 import returnFetch from 'return-fetch';
 
 import type { ArticleFormData, ArticleRequestFormData } from '@/apis/useArticle/article.type';
-import { dateRequestFormat } from '@/libs/utils/dateFormatter';
 
 import API_URL from '../constants/url';
 import interceptor from '../interceptors/interceptor';
+import { formatArticleInitialDataForRequest } from '../utils/formatArticleInitialData';
 
 const options = {
   baseUrl: API_URL.API_BASE_URL,
@@ -32,29 +32,13 @@ const options = {
 const fetchService = returnFetch({ fetch: interceptor.logging(options) });
 
 const ArticleService = {
-  postRegisterArticle: async ({ title, location, date, expense, travelCompanion, travelStyle }: ArticleFormData) => {
+  postRegisterArticle: async (data: ArticleFormData) => {
     try {
-      const formatData: ArticleRequestFormData = {
-        title,
-        location: location.map(({ placeId, address, city }) => ({ place_id: placeId, address, city })),
-        start_at: dateRequestFormat(date.from),
-        end_at: dateRequestFormat(date.to),
-        travel_companion: travelCompanion
-      };
-
-      if (expense) {
-        formatData.expense = String(expense);
-      }
-
-      if (travelStyle.length) {
-        formatData.style = travelStyle;
-      }
-
+      const formatData: ArticleRequestFormData = formatArticleInitialDataForRequest(data);
       const response = await fetchService('api/v1/register/article', {
         method: 'POST',
         body: JSON.stringify(formatData)
       });
-
       const responseData = await response.json();
 
       if (!response.ok) {
